+++ R/query.r
@@ -30,14 +30,17 @@ query_exec <- function(query, project, destination_table = NULL,
                        warn = TRUE,
                        create_disposition = "CREATE_IF_NEEDED",
                        write_disposition = "WRITE_EMPTY",
-                       useLegacySql = TRUE) {
+                       use_legacy_sql = TRUE) {
 
-  dest <- run_query_job(query = query, project = project,
-                        destination_table = destination_table,
-                        default_dataset = default_dataset,
-                        create_disposition = create_disposition,
-                        write_disposition = write_disposition,
-                        useLegacySql = useLegacySql)
+  dest <- run_query_job(
+    query = query,
+    project = project,
+    destination_table = destination_table,
+    default_dataset = default_dataset,
+    create_disposition = create_disposition,
+    write_disposition = write_disposition,
+    use_legacy_sql = use_legacy_sql
+  )
 
   list_tabledata(dest$projectId, dest$datasetId, dest$tableId,
     page_size = page_size, max_pages = max_pages, warn = warn)
@@ -45,17 +48,24 @@ query_exec <- function(query, project, destination_table = NULL,
 
 # Submits a query job, waits for it, and returns information on the destination
 # table for further consumption by the list_tabledata* functions
-run_query_job <- function(query, project, destination_table, default_dataset,
+run_query_job <- function(query,
+                          project,
+                          destination_table,
+                          default_dataset,
                           create_disposition = "CREATE_IF_NEEDED",
                           write_disposition = "WRITE_EMPTY",
-                          useLegacySql = TRUE) {
+                          use_legacy_sql = TRUE) {
   assert_that(is.string(query), is.string(project))
 
-  job <- insert_query_job(query, project, destination_table = destination_table,
-                          default_dataset = default_dataset,
-                          create_disposition = create_disposition,
-                          write_disposition = write_disposition,
-                          useLegacySql = useLegacySql)
+  job <- insert_query_job(
+    query,
+    project,
+    destination_table = destination_table,
+    default_dataset = default_dataset,
+    create_disposition = create_disposition,
+    write_disposition = write_disposition,
+    use_legacy_sql = use_legacy_sql
+  )
   job <- wait_for(job)
 
   job$configuration$query$destinationTable