+++ R/tabledata.r
@@ -37,9 +37,16 @@ list_tabledata <- function(project, dataset, table, page_size = 1e4,
     rows <<- c(rows, list(new_rows))
   }
 
-  list_tabledata_callback(project, dataset, table, append_rows,
-    table_info = table_info, page_size = page_size, max_pages = max_pages,
-    warn = warn, quiet = quiet
+  list_tabledata_callback(
+    project,
+    dataset,
+    table,
+    append_rows,
+    table_info = table_info,
+    page_size = page_size,
+    max_pages = max_pages,
+    warn = warn,
+    quiet = quiet
   )
 
   do.call("rbind", rows)
@@ -159,8 +166,13 @@ list_tabledata_iter <- function(project, dataset, table, table_info = NULL) {
   #' rows using a specified page size), \code{is_complete} (checks if all rows
   #' have been fetched), \code{get_schema} (returns the schema of the table),
   #' and \code{get_rows_fetched} (returns the number of rows already fetched).
-  list(next_ = next_, next_paged = next_paged, is_complete = is_complete,
-       get_schema = get_schema, get_rows_fetched = get_rows_fetched)
+  list(
+    next_ = next_,
+    next_paged = next_paged,
+    is_complete = is_complete,
+    get_schema = get_schema,
+    get_rows_fetched = get_rows_fetched
+  )
 }
 
 #Types can be loaded into R, record is not supported yet.
@@ -169,7 +181,18 @@ converter <- list(
   float = as.double,
   boolean = as.logical,
   string = identity,
-  timestamp = function(x) as.POSIXct(as.integer(x), origin = "1970-01-01", tz = "UTC")
+  timestamp = function(x) {
+    as.POSIXct(as.numeric(x), origin = "1970-01-01", tz = "UTC")
+  },
+  time = function(x) {
+    readr::parse_time(x, "%H:%M:%OS")
+  },
+  date = function(x) {
+    readr::parse_date(x, format = "%Y-%m-%d")
+  },
+  datetime = function(x) {
+    readr::parse_datetime(x)
+  }
 )
 
 extract_data <- function(rows, schema) {